refactor(cli): co-locate event tests with their respective submodules
Move the 1,700-line test block from events/mod.rs into the submodule
files it actually tests (types.rs, filter.rs, listener.rs, broadcaster.rs),
following idiomatic Rust test placement. events/mod.rs is now just
declarations and re-exports.
